Move 3ds logic to Braintree View
