+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/payment/braintree/BraintreePaymentView.kt
@@ -4,8 +4,14 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
+import com.braintreepayments.api.BraintreeFragment
+import com.braintreepayments.api.ThreeDSecure
 import com.braintreepayments.api.dropin.DropInRequest
 import com.braintreepayments.api.dropin.DropInResult
+import com.braintreepayments.api.interfaces.BraintreeErrorListener
+import com.braintreepayments.api.interfaces.PaymentMethodNonceCreatedListener
+import com.braintreepayments.api.models.PaymentMethodNonce
+import com.braintreepayments.api.models.ThreeDSecureRequest
 import com.karhoo.uisdk.screen.booking.booking.payment.PaymentDropInMVP
 import com.karhoo.uisdk.util.extension.isGuest
 
@@ -37,6 +43,33 @@ class BraintreePaymentView : PaymentDropInMVP.View {
         val dropInRequest = DropInRequest().clientToken(braintreeSDKToken)
         val requestCode = if (isGuest()) REQ_CODE_BRAINTREE_GUEST else REQ_CODE_BRAINTREE
         (context as Activity).startActivityForResult(dropInRequest.getIntent(context), requestCode)
+    }
+
+    override fun handleThreeDSecure(context: Context, braintreeSDKToken: String, nonce: String, amount: String) {
+        val braintreeFragment = BraintreeFragment
+                .newInstance(context as AppCompatActivity, braintreeSDKToken)
+
+        braintreeFragment.addListener(object : PaymentMethodNonceCreatedListener {
+            override fun onPaymentMethodNonceCreated(paymentMethodNonce: PaymentMethodNonce?) {
+                actions?.threeDSecureNonce(paymentMethodNonce?.nonce.orEmpty())
+            }
+        })
+
+        braintreeFragment.addListener(
+                object : BraintreeErrorListener {
+                    override fun onError(error: Exception?) {
+                        actions?.showPaymentFailureDialog()
+                    }
+                })
+
+        val threeDSecureRequest = ThreeDSecureRequest()
+                .nonce(nonce)
+                .amount(amount)
+                .versionRequested(ThreeDSecureRequest.VERSION_2)
+        ThreeDSecure.performVerification(braintreeFragment, threeDSecureRequest)
+        { request, lookup ->
+            ThreeDSecure.continuePerformVerification(braintreeFragment, request, lookup)
+        }
     }
 
     companion object {
